input and output mechanism implemented
A small text box that gets the input (people that will be tagged to the post) from the user and an add button that adds this input to the state array have been implemented.
State array that holds all the people that added by the user has also been rendered to allow user to see the added people so far.