clear and remove buttons are added
A clear button that allows users to the clear the selected/added people, and a remove button next to each selected people have been implemented.

Remove button allows user to remove the specific people that they added before. It removes the clicked people element from the state's array on click