state variables and input mechanism implemented
A text box that gets the input (location in which the story have occurred) and an add button that adds this input to the state array have been implemented.
State array has also been rendered to allow user to see the entered locations so far.
State variable named 'value' holds the last location entered by the user.