checkbox and popups are added
A checkbox added. When clicked it allows users to filter posts with the tags content not only with the ones they have entered but also with wiki data of the those.
A pop up added. I pops up when users click on a tag. It shows the contents of the wiki data of that tag.
Remark: These mechanisms haven't integrated with backend end points yet.